server refactored to be a module and receive data in config
